Release 0.11.0

+++ src/analyzer/analyzers/analyzer_crystal_kemal.cr
@@ -1,18 +1,18 @@
 require "../../models/analyzer"
 
-class AnalyzerSinatra < Analyzer
+class AnalyzerCrystalKemal < Analyzer
   def analyze
     # Source Analysis
     begin
       Dir.glob("#{@base_path}/**/*") do |path|
         next if File.directory?(path)
-        if File.exists?(path)
+        if File.exists?(path) && File.extname(path) == ".cr" && !path.includes?("lib")
           File.open(path, "r", encoding: "utf-8", invalid: :skip) do |file|
             last_endpoint = Endpoint.new("", "")
             file.each_line do |line|
               endpoint = line_to_endpoint(line)
               if endpoint.method != ""
-                @result << endpoint
+                result << endpoint
                 last_endpoint = endpoint
               end
 
@@ -30,28 +30,38 @@ class AnalyzerSinatra < Analyzer
       logger.debug e
     end
 
-    @result
+    result
   end
 
   def line_to_param(content : String) : Param
-    if content.includes? "param["
-      param = content.split("param[")[1].split("]")[0].gsub("\"", "").gsub("'", "")
+    if content.includes? "env.params.query["
+      param = content.split("env.params.query[")[1].split("]")[0].gsub("\"", "").gsub("'", "")
       return Param.new(param, "", "query")
     end
 
-    if content.includes? "params["
-      param = content.split("params[")[1].split("]")[0].gsub("\"", "").gsub("'", "")
-      return Param.new(param, "", "query")
+    if content.includes? "env.params.json["
+      param = content.split("env.params.json[")[1].split("]")[0].gsub("\"", "").gsub("'", "")
+      return Param.new(param, "", "json")
     end
 
-    if content.includes? "request.env["
-      param = content.split("request.env[")[1].split("]")[0].gsub("\"", "").gsub("'", "")
+    if content.includes? "env.params.body["
+      param = content.split("env.params.body[")[1].split("]")[0].gsub("\"", "").gsub("'", "")
+      return Param.new(param, "", "form")
+    end
+
+    if content.includes? "env.request.headers["
+      param = content.split("env.request.headers[")[1].split("]")[0].gsub("\"", "").gsub("'", "")
       return Param.new(param, "", "header")
     end
 
-    if content.includes? "headers["
-      param = content.split("headers[")[1].split("]")[0].gsub("\"", "").gsub("'", "")
-      return Param.new(param, "", "header")
+    if content.includes? "env.request.cookies["
+      param = content.split("env.request.cookies[")[1].split("]")[0].gsub("\"", "").gsub("'", "")
+      return Param.new(param, "", "cookie")
+    end
+
+    if content.includes? "cookies.get_raw("
+      param = content.split("cookies.get_raw(")[1].split(")")[0].gsub("\"", "").gsub("'", "")
+      return Param.new(param, "", "cookie")
     end
 
     Param.new("", "", "")
@@ -100,11 +110,19 @@ class AnalyzerSinatra < Analyzer
       end
     end
 
+    content.scan(/ws\s+['"](.+?)['"]/) do |match|
+      if match.size > 1
+        endpoint = Endpoint.new("#{@url}#{match[1]}", "GET")
+        endpoint.set_protocol("ws")
+        return endpoint
+      end
+    end
+
     Endpoint.new("", "")
   end
 end
 
-def analyzer_sinatra(options : Hash(Symbol, String))
-  instance = AnalyzerSinatra.new(options)
+def analyzer_crystal_kemal(options : Hash(Symbol, String))
+  instance = AnalyzerCrystalKemal.new(options)
   instance.analyze
 end
